스타일링 & 14강까지 기능추가

+++ src/controllers/videoController.js
@@ -1,35 +1,41 @@
-import { redirect } from "express/lib/response";
+import User from "../models/User";
 import Video from "../models/Video";
 
 export const home = async(req, res) => {
     try{
-        const videos = await Video.find({});
+        const videos = await Video.find({}).sort({createdAt: "desc"});
         return res.render("home", {pageTitle : "Home", videos})
     } catch(error) {
-        return res.render("server-error", {error}); 
+        return res.status(404).render("server-error", {error}); 
     }
 
 };
 export const watch = async(req, res) => {
     const {id} = req.params;
-    const video = await Video.findById(id);
+    const video = await Video.findById(id).populate("owner");
     if(!video) {
         return res.render("404", {pageTitle: "Video not found"});
     }
-    return res.render("watch", {pageTitle :video.title, video});
+    return res.render("videos/watch", {pageTitle :video.title, video});
 };
 export const getEdit = async(req, res) => {
     const {id} = req.params;
+    const {user: {_id}} =req.session;
     const video = await Video.findById(id);
 
     if(!video) {
         return res.render("404", {pageTitle: "Video not found"});
     }
-    return res.render("edit", {pageTitle: `editing ${video.title}`, video})
+
+    if(String(video.owner) !== String(_id)) {
+        return res.status(403).redirect("/");
+    }
+    return res.render("videos/edit", {pageTitle: `editing ${video.title}`, video})
 };
 
 export const postEdit = async(req, res) => {
     const {id} = req.params;
+    const {user: {_id}} =req.session;
     const video = await Video.exists({_id: id});
     const {title, description, hashtags} = req.body;
 
@@ -37,36 +43,86 @@ export const postEdit = async(req, res) => {
         return res.render("404", {pageTitle: "Video not found"});
     }
 
+    if(String(video.owner) !== String(_id)) {
+        return res.status(403).redirect("/");
+    }
+
     await Video.findByIdAndUpdate(id, {
         title,
         description,
         hashtags : Video.formatHashtags(hashtags),
     })
-    console.log(hashtags);
     res.redirect(`/videos/${id}`);
 }
 
 export const getUpload = (req, res) => {
-    return res.render("upload", {pageTitle: "upload video"})
+    return res.render("videos/upload", {pageTitle: "upload video"})
 }
 
 export const postUpload = async(req, res) => {
+    const {user: {
+        _id
+    } } = req.session;
+    const {video, thumb} = req.files;
     const {title, description, hashtags} = req.body;
     try {
-        await Video.create({
+        const newVideo = await Video.create({
             title,
             description,
+            fileUrl:video[0].path,
+            thumbUrl: thumb[0].path,
+            owner: _id,
             hashtags : Video.formatHashtags(hashtags),
         })
+        const user = await User.findById(_id);
+        user.videos.push(newVideo._id);
+        user.save();
+        return res.redirect("/");
     } catch(error) {
-        return res.render("upload", {pageTitle: "upload video", errorMessage: error._message});
+        return res.status(404).render("upload", {pageTitle: "upload video", errorMessage: error._message});
     }
-    return res.redirect("/");
+
 }
 
 
 export const deleteVideo = async(req, res) => {
     const {id} = req.params;
+    const { user: {_id}} = req.session;
+    const video = await Video.findById(_id);
+    if(!video) {
+        return res.status(400).render("404", {pageTitle:"Video not found"});
+    }
+    
+    if(String(video.owner) !== String(_id)) {
+        return res.status(403).redirect("/");
+    }
     await Video.findByIdAndDelete(id); //findDelete를 사용, remove 는 롤백이 되지 않음
+
     return res.redirect("/");
 };
+
+export const search = async(req, res) => {
+    const {keyword} = req.query;
+    let videos = [];
+    if(keyword) {
+        videos = await Video.find({
+            title: {
+                $regex: new RegExp(keyword, "i")
+            },
+        })
+    } 
+    console.log("serach");
+    return res.render("videos/search", {pageTitle: "search", videos});
+}
+
+export const registerView = async(req, res) => {
+    const {id} = req.params;
+    const video = await Video.findById(id);
+
+    if(!video) {
+        return res.sendStatus(404);
+    }
+    video.meta.views = video.meta.views + 1;
+    await video.save();
+    return res.sendStatus(200)
+}
